backend/requirements.txt
numpy
fastapi
uvicorn
python-dotenv
pandas
google-genai